Migrate forgot password, support custom middleware for endpoints
