+++ server/service/handler.go
@@ -24,12 +24,11 @@ import (
 
 // FleetEndpoints is a collection of RPC endpoints implemented by the Fleet API.
 type FleetEndpoints struct {
-	Login          endpoint.Endpoint
-	Logout         endpoint.Endpoint
-	ForgotPassword endpoint.Endpoint
-	InitiateSSO    endpoint.Endpoint
-	CallbackSSO    endpoint.Endpoint
-	SSOSettings    endpoint.Endpoint
+	Login       endpoint.Endpoint
+	Logout      endpoint.Endpoint
+	InitiateSSO endpoint.Endpoint
+	CallbackSSO endpoint.Endpoint
+	SSOSettings endpoint.Endpoint
 }
 
 // MakeFleetServerEndpoints creates the Fleet API endpoints.
@@ -41,11 +40,7 @@ func MakeFleetServerEndpoints(svc fleet.Service, urlPrefix string, limitStore th
 			throttled.RateQuota{MaxRate: throttled.PerMin(10), MaxBurst: 9})(
 			makeLoginEndpoint(svc),
 		),
-		Logout: logged(makeLogoutEndpoint(svc)),
-		ForgotPassword: limiter.Limit(
-			throttled.RateQuota{MaxRate: throttled.PerHour(10), MaxBurst: 9})(
-			logged(makeForgotPasswordEndpoint(svc)),
-		),
+		Logout:      logged(makeLogoutEndpoint(svc)),
 		InitiateSSO: logged(makeInitiateSSOEndpoint(svc)),
 		CallbackSSO: logged(makeCallbackSSOEndpoint(svc, urlPrefix)),
 		SSOSettings: logged(makeSSOSettingsEndpoint(svc)),
@@ -53,12 +48,11 @@ func MakeFleetServerEndpoints(svc fleet.Service, urlPrefix string, limitStore th
 }
 
 type fleetHandlers struct {
-	Login          http.Handler
-	Logout         http.Handler
-	ForgotPassword http.Handler
-	InitiateSSO    http.Handler
-	CallbackSSO    http.Handler
-	SettingsSSO    http.Handler
+	Login       http.Handler
+	Logout      http.Handler
+	InitiateSSO http.Handler
+	CallbackSSO http.Handler
+	SettingsSSO http.Handler
 }
 
 func makeKitHandlers(e FleetEndpoints, opts []kithttp.ServerOption) *fleetHandlers {
@@ -67,12 +61,11 @@ func makeKitHandlers(e FleetEndpoints, opts []kithttp.ServerOption) *fleetHandle
 		return kithttp.NewServer(e, decodeFn, encodeResponse, opts...)
 	}
 	return &fleetHandlers{
-		Login:          newServer(e.Login, decodeLoginRequest),
-		Logout:         newServer(e.Logout, decodeNoParamsRequest),
-		ForgotPassword: newServer(e.ForgotPassword, decodeForgotPasswordRequest),
-		InitiateSSO:    newServer(e.InitiateSSO, decodeInitiateSSORequest),
-		CallbackSSO:    newServer(e.CallbackSSO, decodeCallbackSSORequest),
-		SettingsSSO:    newServer(e.SSOSettings, decodeNoParamsRequest),
+		Login:       newServer(e.Login, decodeLoginRequest),
+		Logout:      newServer(e.Logout, decodeNoParamsRequest),
+		InitiateSSO: newServer(e.InitiateSSO, decodeInitiateSSORequest),
+		CallbackSSO: newServer(e.CallbackSSO, decodeCallbackSSORequest),
+		SettingsSSO: newServer(e.SSOSettings, decodeNoParamsRequest),
 	}
 }
 
@@ -153,7 +146,7 @@ func MakeHandler(svc fleet.Service, config config.FleetConfig, logger kitlog.Log
 	}
 
 	attachFleetAPIRoutes(r, fleetHandlers)
-	attachNewStyleFleetAPIRoutes(r, svc, logger, fleetAPIOptions)
+	attachNewStyleFleetAPIRoutes(r, svc, logger, limitStore, fleetAPIOptions)
 
 	// Results endpoint is handled different due to websockets use
 	// TODO: this would probably not work once v1 is deprecated
@@ -249,13 +242,12 @@ func addMetrics(r *mux.Router) {
 func attachFleetAPIRoutes(r *mux.Router, h *fleetHandlers) {
 	r.Handle("/api/v1/fleet/login", h.Login).Methods("POST").Name("login")
 	r.Handle("/api/v1/fleet/logout", h.Logout).Methods("POST").Name("logout")
-	r.Handle("/api/v1/fleet/forgot_password", h.ForgotPassword).Methods("POST").Name("forgot_password")
 	r.Handle("/api/v1/fleet/sso", h.InitiateSSO).Methods("POST").Name("intiate_sso")
 	r.Handle("/api/v1/fleet/sso", h.SettingsSSO).Methods("GET").Name("sso_config")
 	r.Handle("/api/v1/fleet/sso/callback", h.CallbackSSO).Methods("POST").Name("callback_sso")
 }
 
-func attachNewStyleFleetAPIRoutes(r *mux.Router, svc fleet.Service, logger kitlog.Logger, opts []kithttp.ServerOption) {
+func attachNewStyleFleetAPIRoutes(r *mux.Router, svc fleet.Service, logger kitlog.Logger, limitStore throttled.GCRAStore, opts []kithttp.ServerOption) {
 	// user-authenticated endpoints
 	ue := newUserAuthenticatedEndpointer(svc, opts, r, "v1")
 
@@ -419,7 +411,12 @@ func attachNewStyleFleetAPIRoutes(r *mux.Router, svc fleet.Service, logger kitlo
 	ne.POST("/api/_version_/fleet/perform_required_password_reset", performRequiredPasswordResetEndpoint, performRequiredPasswordResetRequest{})
 	ne.POST("/api/_version_/fleet/users", createUserFromInviteEndpoint, createUserRequest{})
 	ne.GET("/api/_version_/fleet/invites/{token}", verifyInviteEndpoint, verifyInviteRequest{})
-	ne.POST("/api/v1/fleet/reset_password", resetPasswordEndpoint, resetPasswordRequest{})
+	ne.POST("/api/_version_/fleet/reset_password", resetPasswordEndpoint, resetPasswordRequest{})
+
+	limiter := ratelimit.NewMiddleware(limitStore)
+	ne.WithCustomMiddleware(
+		limiter.Limit(throttled.RateQuota{MaxRate: throttled.PerHour(10), MaxBurst: 9})).
+		POST("/api/v1/fleet/forgot_password", forgotPasswordEndpoint, forgotPasswordRequest{})
 }
 
 // TODO: this duplicates the one in makeKitHandler
