+++ server/service/handler_test.go
@@ -42,10 +42,6 @@ func TestAPIRoutes(t *testing.T) {
 			verb: "POST",
 			uri:  "/api/v1/fleet/login",
 		},
-		{
-			verb: "POST",
-			uri:  "/api/v1/fleet/forgot_password",
-		},
 	}
 
 	for _, route := range routes {
