+++ server/service/users.go
@@ -6,6 +6,8 @@ import (
 	"encoding/base64"
 	"errors"
 	"html/template"
+	"net/http"
+	"time"
 
 	"github.com/fleetdm/fleet/v4/server"
 	"github.com/fleetdm/fleet/v4/server/authz"
@@ -840,4 +842,82 @@ func (svc *Service) ResetPassword(ctx context.Context, token, password string) e
 	}
 
 	return nil
+}
+
+////////////////////////////////////////////////////////////////////////////////
+// Forgot Password
+////////////////////////////////////////////////////////////////////////////////
+
+type forgotPasswordRequest struct {
+	Email string `json:"email"`
+}
+
+type forgotPasswordResponse struct {
+	Err error `json:"error,omitempty"`
+}
+
+func (r forgotPasswordResponse) error() error { return r.Err }
+func (r forgotPasswordResponse) status() int  { return http.StatusAccepted }
+
+func forgotPasswordEndpoint(ctx context.Context, request interface{}, svc fleet.Service) (interface{}, error) {
+	req := request.(*forgotPasswordRequest)
+	// Any error returned by the service should not be returned to the
+	// client to prevent information disclosure (it will be logged in the
+	// server logs).
+	_ = svc.RequestPasswordReset(ctx, req.Email)
+	return forgotPasswordResponse{}, nil
+}
+
+func (svc *Service) RequestPasswordReset(ctx context.Context, email string) error {
+	// skipauth: No viewer context available. The user is locked out of their
+	// account and trying to reset their password.
+	svc.authz.SkipAuthorization(ctx)
+
+	// Regardless of error, sleep until the request has taken at least 1 second.
+	// This means that any request to this method will take ~1s and frustrate a timing attack.
+	defer func(start time.Time) {
+		time.Sleep(time.Until(start.Add(1 * time.Second)))
+	}(time.Now())
+
+	user, err := svc.ds.UserByEmail(ctx, email)
+	if err != nil {
+		return err
+	}
+	if user.SSOEnabled {
+		return ctxerr.New(ctx, "password reset for single sign on user not allowed")
+	}
+
+	random, err := server.GenerateRandomText(svc.config.App.TokenKeySize)
+	if err != nil {
+		return err
+	}
+	token := base64.URLEncoding.EncodeToString([]byte(random))
+
+	request := &fleet.PasswordResetRequest{
+		ExpiresAt: time.Now().Add(time.Hour * 24),
+		UserID:    user.ID,
+		Token:     token,
+	}
+	_, err = svc.ds.NewPasswordResetRequest(ctx, request)
+	if err != nil {
+		return err
+	}
+
+	config, err := svc.ds.AppConfig(ctx)
+	if err != nil {
+		return err
+	}
+
+	resetEmail := fleet.Email{
+		Subject: "Reset Your Fleet Password",
+		To:      []string{user.Email},
+		Config:  config,
+		Mailer: &mail.PasswordResetMailer{
+			BaseURL:  template.URL(config.ServerSettings.ServerURL + svc.config.Server.URLPrefix),
+			AssetURL: getAssetURL(),
+			Token:    token,
+		},
+	}
+
+	return svc.mailService.SendEmail(resetEmail)
 }
