+++ server/service/users_test.go
@@ -2,6 +2,7 @@ package service
 
 import (
 	"context"
+	"database/sql"
 	"errors"
 	"testing"
 	"time"
@@ -593,4 +594,69 @@ func TestPerformRequiredPasswordReset(t *testing.T) {
 			assert.False(t, u.AdminForcedPasswordReset)
 		})
 	}
+}
+
+func TestResetPassword(t *testing.T) {
+	ds := mysql.CreateMySQLDS(t)
+
+	svc := newTestService(ds, nil, nil)
+	createTestUsers(t, ds)
+	passwordResetTests := []struct {
+		token       string
+		newPassword string
+		wantErr     error
+	}{
+		{ // all good
+			token:       "abcd",
+			newPassword: "123cat!",
+		},
+		{ // prevent reuse
+			token:       "abcd",
+			newPassword: "123cat!",
+			wantErr:     fleet.NewInvalidArgumentError("new_password", "cannot reuse old password"),
+		},
+		{ // bad token
+			token:       "dcbaz",
+			newPassword: "123cat!",
+			wantErr:     sql.ErrNoRows,
+		},
+		{ // missing token
+			newPassword: "123cat!",
+			wantErr:     fleet.NewInvalidArgumentError("token", "Token cannot be empty field"),
+		},
+	}
+
+	for _, tt := range passwordResetTests {
+		t.Run("", func(t *testing.T) {
+			request := &fleet.PasswordResetRequest{
+				UpdateCreateTimestamps: fleet.UpdateCreateTimestamps{
+					CreateTimestamp: fleet.CreateTimestamp{
+						CreatedAt: time.Now(),
+					},
+					UpdateTimestamp: fleet.UpdateTimestamp{
+						UpdatedAt: time.Now(),
+					},
+				},
+				ExpiresAt: time.Now().Add(time.Hour * 24),
+				UserID:    1,
+				Token:     "abcd",
+			}
+			_, err := ds.NewPasswordResetRequest(context.Background(), request)
+			assert.Nil(t, err)
+
+			serr := svc.ResetPassword(test.UserContext(&fleet.User{ID: 1}), tt.token, tt.newPassword)
+			if tt.wantErr != nil {
+				assert.Equal(t, tt.wantErr.Error(), ctxerr.Cause(serr).Error())
+			} else {
+				assert.Nil(t, serr)
+			}
+		})
+	}
+}
+
+func refreshCtx(t *testing.T, ctx context.Context, user *fleet.User, ds fleet.Datastore, session *fleet.Session) context.Context {
+	reloadedUser, err := ds.UserByEmail(ctx, user.Email)
+	require.NoError(t, err)
+
+	return viewer.NewContext(ctx, viewer.Viewer{User: reloadedUser, Session: session})
 }
